# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AGORA Core API",
    "description": "Agent Grid Open Runtime Architecture — AI Service Hub Core. Agent identity, wallet, service registry, and execution.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/admin/instance/{id}/status": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "description": "Update instance status (e.g. flag, block, unblock). Setting to non-registered suspends exported services.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "unregistered",
                      "pending",
                      "registered",
                      "flagged",
                      "blocked"
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/instance/register": {
      "post": {
        "tags": [
          "Instance"
        ],
        "description": "Register this installation. Returns registration_code (shown once).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "owner_email"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "owner_email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "instance_id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "status": {
                          "type": "string"
                        },
                        "registration_code": {
                          "type": "string"
                        },
                        "expires_at": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/instance/activate": {
      "post": {
        "tags": [
          "Instance"
        ],
        "description": "Activate instance with registration_code and activation_token (from Official or admin).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "instance_id",
                  "registration_code",
                  "activation_token"
                ],
                "properties": {
                  "instance_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "registration_code": {
                    "type": "string"
                  },
                  "activation_token": {
                    "type": "string"
                  },
                  "official_issuer_id": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "instance_id": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "registered_at": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
...
}